pyproject.toml
[project]
name = "kia-bot"
version = "0.1.0"
description = "kia-bot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.61.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-orm>=0.8.3",
]

[project.optional-dependencies]
dev = [
    "pyright[nodejs]"
]

[project.dependency-groups]
dev = [
    "pyright",
]
[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]
nonebot-plugin-localstore = ["nonebot_plugin_localstore"]
nonebot-plugin-orm = ["nonebot_plugin_orm"]


[tool.pyright]
# For Pylance/Pyright configurations, see https://microsoft.github.io/pyright/#/configuration.
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "standard"
